cysco_api/requirements.txt
certifi==2021.5.30
chardet==4.0.0
confluent-kafka==1.7.0
fastavro==1.4.1
idna==2.10
psycopg2-binary==2.8.6
python-dotenv==0.17.1
requests==2.25.1
urllib3==1.26.5
